First pass at this.
* add gibberish config
* remove old tasks

I'd like to leave 'test' as a custom task.  Right now the declaration of
js builds using jshint is a bit of a hack.  It would be nice to be able
to declare the hinting directly.  Only the single file needs to be built
(and then, it only needs to be compiled).  It might be nice to
declaratively configure the watchers for connect.reload() as well.